DÄ SÄNASTE

+++ trunk/src/se/combitech/strokesformartians/dancing/IMartianMesh.java
@@ -6,5 +6,5 @@ public interface IMartianMesh
 	
 	public float[] getTextureCoordinates() throws Exception;
 	
-	public byte[] getIndices() throws Exception;
+	public short[] getIndices() throws Exception;
 }
